Add index.jsx into each component folder
