+++ src/components/app/index.jsx
@@ -1,0 +1,3 @@
+import App from './app';
+
+export default App;
