+++ src/components/card-list-item/card-list-item.css
@@ -1,0 +1,4 @@
+.card-list-item {
+  cursor: pointer;
+  user-select: none;
+}
