+++ src/components/card-list-item/index.jsx
@@ -1,0 +1,3 @@
+import CardListItem from './card-list-item';
+
+export default CardListItem;
