+++ src/components/card-list/index.jsx
@@ -1,0 +1,3 @@
+import CardList from './card-list';
+
+export default CardList;
